use exception e to pass pre-commit
