+++ src/orca_hls_utils/DateRangeHLSStream.py
@@ -145,7 +145,26 @@ class DateRangeHLSStream:
         stream_url = "{}/hls/{}/live.m3u8".format(
             (self.stream_base), (current_folder)
         )
-        stream_obj = m3u8.load(stream_url)
+
+        try:
+            stream_obj = m3u8.load(stream_url)
+        except Exception as e:
+            print(f"Failed to load m3u8 playlist from {stream_url}: {e}")
+            # Move to next folder or end stream
+            if self.current_folder_index + 1 >= len(self.valid_folders):
+                self.is_end_of_stream = True
+                return None, None, None
+
+            self.current_folder_index += 1
+            if self.current_folder_index >= len(self.valid_folders):
+                self.is_end_of_stream = True
+                return None, None, None
+
+            self.current_clip_start_time = int(
+                self.valid_folders[self.current_folder_index]
+            )
+            return None, None, None
+
         num_total_segments = len(stream_obj.segments)
         print(num_total_segments)
